% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCDBcurator.R
\name{CCDBcurator}
\alias{CCDBcurator}
\title{Cleans all chromosome records}
\usage{
CCDBcurator(cleanCCDBsample)
}
\arguments{
\item{cleanCCDBsample}{Input is a sample of CCDB rows downloaded by chromer}
}
\description{
Cleans all chromosome records
}
\details{
CCDBcurator cleans all gametophytic and sporophytic counts of CCDB dataset (downloaded by chromer R package). Creates a dataframe with haploid chromsome number and possible aneuploidy column
}
