% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sporophytictranslate.R
\name{sporophytic.translator}
\alias{sporophytic.translator}
\title{Cleans sporophytic counts}
\usage{
sporophytic.translator(rowfromCCDB)
}
\arguments{
\item{rowfromCCDB}{Input is a row of CCDB downloaded by chromer}
}
\value{
A data frame with 1 row per chromosome number entry cleaned and the original record
}
\description{
Cleans sporophytic counts
}
\details{
sporphytic.translator cleans the sporophytic counts of CCDB dataset (downloaded by chromer R package)
}
